project/requirements.txt
aiofiles==23.1.0
celery==5.2.7
fastapi[all]
flower==1.2.0
Jinja2==3.1.2
pytest==7.2.2
redis==4.5.4
requests==2.28.2
uvicorn==0.21.1
httpx==0.23.3
Pillow
PyJWT
passlib
python-jose[cryptography]
bcrypt
databases[postgresql]
python-dotenv
psycopg2-binary